// src/app/interfaces/program.interface.ts
export interface PostgraduateProgram {
  program: string;
  session: string;
  applicationLastDate: string;
  admissionTestDate: string;
  status: string;
}

export interface UndergraduateInfo {
  academicSession: string;
  applicationDeadline: string;
  admissionTest: string;
  resultPublish: string;
  registration: string;
}